refactor: split use reading history
