+++ packages/webapp/pages/history.tsx
@@ -3,6 +3,7 @@ import classNames from 'classnames';
 import { NextSeo } from 'next-seo';
 import { ResponsiveNoPaddingPageContainer } from '@dailydotdev/shared/src/components/utilities';
 import useReadingHistory from '@dailydotdev/shared/src/hooks/useReadingHistory';
+import useInfiniteReadingHistory from '@dailydotdev/shared/src/hooks/useInfiniteReadingHistory';
 import ReadingHistoryList from '@dailydotdev/shared/src/components/history/ReadingHistoryList';
 import ReadingHistoryPlaceholder from '@dailydotdev/shared/src/components/history/ReadingHistoryPlaceholder';
 import ReadingHistoryEmptyScreen from '@dailydotdev/shared/src/components/history/ReadingHistoryEmptyScreen';
@@ -11,20 +12,16 @@ import { useRouter } from 'next/router';
 import { getLayout } from '../components/layouts/MainLayout';
 
 const History = (): ReactElement => {
+  const seo = <NextSeo title="Reading History" nofollow noindex />;
   const router = useRouter();
   const { user, tokenRefreshed } = useContext(AuthContext);
-  const {
-    data,
-    isInitialLoading,
-    isLoading,
-    hasData,
-    infiniteScrollRef,
-    hideReadHistory,
-  } = useReadingHistory(user);
-  const seo = <NextSeo title="Reading History" nofollow noindex />;
+  const key = ['readHistory', user?.id];
+  const { hideReadHistory } = useReadingHistory(key);
+  const { data, isInitialLoading, isLoading, hasData, infiniteScrollRef } =
+    useInfiniteReadingHistory(key);
 
   useEffect(() => {
-    if (!user && tokenRefreshed) {
+    if (tokenRefreshed && !user) {
       router.replace('/');
     }
   }, [tokenRefreshed, user]);
